requirements.txt
streamlit==1.32.0
streamlit-webrtc==0.47.9
mediapipe==0.10.11
opencv-python-headless==4.9.0.80
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.3.2
google-genai==0.3.0
python-dotenv==1.0.1
av==10.0.0